feat: Add demo login hint
