+++ src/features/auth/LoginForm.tsx
@@ -1,57 +1,15 @@
 import React from 'react';
 
-import {
-  Alert,
-  AlertDescription,
-  Box,
-  BoxProps,
-  Button,
-  Flex,
-  Stack,
-} from '@chakra-ui/react';
+import { Box, BoxProps, Button, Flex, Stack } from '@chakra-ui/react';
 import { Formiz, useForm } from '@formiz/core';
 import { useQueryClient } from '@tanstack/react-query';
-import { Trans, useTranslation } from 'react-i18next';
+import { useTranslation } from 'react-i18next';
 import { Link } from 'react-router-dom';
 
 import { FieldInput } from '@/components/FieldInput';
 import { useToastError } from '@/components/Toast';
+import { DemoModeHint } from '@/features/auth/DemoModeHint';
 import { useLogin } from '@/features/auth/service';
-
-const MockedApiHint = () => {
-  const { t } = useTranslation(['auth']);
-  const form = useForm({ subscribe: 'form' });
-  const mockedUsername = 'admin';
-  const mockedPassword = 'admin';
-
-  if (process.env.NEXT_PUBLIC_API_BASE_URL) return null;
-
-  return (
-    <Alert mt="4" borderRadius="md" textAlign="center" colorScheme="brand">
-      <AlertDescription>
-        <Trans
-          t={t}
-          i18nKey="auth:mockedApi.loginHint"
-          values={{ credentials: `${mockedUsername}/${mockedPassword}` }}
-          components={{
-            button: (
-              <Button
-                variant="link"
-                color="inherit"
-                onClick={() =>
-                  form.setFieldsValues({
-                    username: mockedUsername,
-                    password: mockedPassword,
-                  })
-                }
-              />
-            ),
-          }}
-        />
-      </AlertDescription>
-    </Alert>
-  );
-};
 
 type LoginFormProps = BoxProps & { onSuccess?: () => void };
 
@@ -118,7 +76,7 @@ export const LoginForm = ({
             </Button>
           </Flex>
 
-          <MockedApiHint />
+          <DemoModeHint />
         </Stack>
       </Formiz>
     </Box>
