+++ src/locales/ar/auth.json
@@ -27,7 +27,7 @@
       "required": "كلمة المرور مطلوبة"
     }
   },
-  "mockedApi": {
-    "loginHint": "الإتصال بواجهة API وهمي، يمكنك إستعمال  <button> {{credentials}} </button> لتسجيل الدخول."
+  "demoMode": {
+    "loginHint": "هذا عرض توضيحي للقراءة فقط ، يمكنك استخدام <button>{{credentials}}</button> لتسجيل الدخول."
   }
 }
