+++ src/locales/en/auth.json
@@ -27,7 +27,7 @@
       "required": "Password is required"
     }
   },
-  "mockedApi": {
-    "loginHint": "API connexion is mocked, you can use <button>{{credentials}}</button> to log in."
+  "demoMode": {
+    "loginHint": "This is a read-only demo, you can use <button>{{credentials}}</button> to login."
   }
 }
