+++ src/locales/fr/auth.json
@@ -27,7 +27,7 @@
       "required": "Le mot de passe est requis"
     }
   },
-  "mockedApi": {
-    "loginHint": "La connexion API est simulée, vous pouvez utiliser <button>{{credentials}}</button> pour vous connecter."
+  "demoMode": {
+    "loginHint": "Ceci est une démo en lecture seule, vous pouvez utiliser <button>{{credentials}}</button> pour vous connecter."
   }
 }
