+++ src/locales/sw/auth.json
@@ -27,7 +27,7 @@
       "required": "Nenosiri linahitajika"
     }
   },
-  "mockedApi": {
-    "loginHint": "Kuingia kwa API kumeigwa, unaweza kutumia <button>{{credentials}}</button> kuingia."
+  "demoMode": {
+    "loginHint": "Hili ni onyesho la kusoma pekee, unaweza kutumia <button>{{credentials}}</button> kuingia."
   }
 }
